New folder (2)/requirements.txt
Flask>=2.2
gunicorn>=20.1.0
psycopg2-binary>=2.9.6
SQLAlchemy>=1.4
Werkzeug>=2.2
